Refactor home page file
